update book<->author to ManyToMany update tests update default data set

+++ library/src/main/java/ru/otus/spring/library/service/impl/AuthorServiceImpl.java
@@ -2,7 +2,9 @@ package ru.otus.spring.library.service.impl;
 
 import org.springframework.stereotype.Service;
 import ru.otus.spring.library.domain.Author;
+import ru.otus.spring.library.domain.Book;
 import ru.otus.spring.library.repository.AuthorDao;
+import ru.otus.spring.library.repository.BookDao;
 import ru.otus.spring.library.repository.JpaRepositoryException;
 import ru.otus.spring.library.service.AuthorService;
 import ru.otus.spring.library.service.LibraryServiceException;
@@ -14,10 +16,12 @@ import java.util.Optional;
 @Service
 public class AuthorServiceImpl implements AuthorService {
     private final AuthorDao authorDao;
+    private final BookDao bookDao;
     private final MessageService ms;
 
-    public AuthorServiceImpl(AuthorDao authorDao, MessageService ms) {
+    public AuthorServiceImpl(AuthorDao authorDao, BookDao bookDao, MessageService ms) {
         this.authorDao = authorDao;
+        this.bookDao = bookDao;
         this.ms = ms;
     }
 
@@ -52,6 +56,10 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public void deleteAuthorById(long id) throws LibraryServiceException {
+        List<Book> allBooksWhereAuthorId = bookDao.getAllBooksWhereAuthorId(id);
+        if (allBooksWhereAuthorId.size() > 0) {
+            throw new LibraryServiceException("Can't delete author with id: " + id + ". Author still has books!");
+        }
         try {
             authorDao.deleteAuthorById(id);
         } catch (JpaRepositoryException e) {
